+++ amazon/dynamodb/fake/src/main/kotlin/org/http4k/connect/amazon/dynamodb/grammar/AttributeExists.kt
@@ -1,0 +1,20 @@
+package org.http4k.connect.amazon.dynamodb.grammar
+
+import org.http4k.connect.amazon.dynamodb.model.AttributeName
+import parser4k.Parser
+import parser4k.commonparsers.Tokens
+import parser4k.commonparsers.token
+import parser4k.inOrder
+import parser4k.map
+import parser4k.skipWrapper
+
+object AttributeExists : ExprFactory {
+    override operator fun invoke(parser: () -> Parser<Expr>) =
+        inOrder(token("attribute_exists"), token("("), Tokens.identifier, token(")"))
+            .skipWrapper()
+            .map { (_, name) ->
+                Expr {
+                    it.item.containsKey(AttributeName.of(name))
+                }
+            }
+}